lom: adding new filters (honor and elite)
