+++ lom/js/comp-filters.js
@@ -1,0 +1,21 @@
+export const filters = `
+<div class="nav">
+  <div class="controllers">
+    <button id="all">ALL</button>
+    <button id="amen">AMEN</button>
+    <button id="es">ES</button>
+    <button id="espt">ESPT</button>
+    <button id="pt">PT</button>
+    <button id="euen">EUEN</button>
+    <button id="mush">MUSH</button>
+    <button id="de">DE</button>
+    <button id="fr">FR</button>
+    <button id="me">ME</button>
+    <button id="tr">TR</button>
+    <button id="ru">RU</button>
+    <button id="honor">HONOR</button>
+    <button id="elite">ELITE</button>
+    <button id="top">TOP 100</button>
+  </div>
+</div>
+`
