+++ lom/js/db-merges.js
@@ -1,4 +1,4 @@
-const merges = {
+export const merges = {
   '2024-07-08': {
     1001: [1088],
     1003: [1036],
@@ -942,403 +942,3 @@ const merges = {
 //   42121: [42122, 42123],
 //   42124: [42125, 42126],
 // }
-
-class User {
-  constructor(name, server, id) {
-    this.name = name.join('/')
-    this.server = server
-    this.id = id
-  }
-}
-
-class UserTop extends User {
-  constructor(name, server, id) {
-    super(name, server, id)
-    this.group = ['top']
-  }
-}
-
-class UserHonorMention extends User {
-  constructor(name, server, id) {
-    super(name, server, id)
-    this.group = ['honor']
-  }
-}
-
-class UserMestre extends User {
-  constructor(name, server, id) {
-    super(name, server, id)
-    this.group = ['mestre']
-  }
-}
-
-const friends = [
-  new User(['DKiingG7'], [11042], null, ['friend']),
-  new User(['AhkTaar'], [30001], null, ['friend']),
-  // new User(['Talisa'], [1310, 6103], null, ['friend']),
-  // new User(['Nico'], [1310, 6103], null, ['friend']),
-  // new User(['Doruk'], [11174], null, ['friend']),
-  // new User(['Koturno'], [1388, 11220, 11222], null, ['friend']),
-  // new User(['Neo'], [11256], null, ['friend']),
-  // new User(['Diego'], [11257], null, ['friend']),
-  // new User(['Gustavo'], [11310], null, ['friend']),
-  // new User(['Murilo'], [11300], null, ['friend']),
-  // new User(['Cleo'], [11323], null, ['friend']),
-  // new User(['Max'], [11323], null, ['friend']),
-  // new User(['Mandy'], [11323], null, ['friend']),
-  // new User(['Magneto'], [11323], null, ['friend']),
-  // new User(['Ipek'], [36264], null, ['friend']),
-]
-
-const topUsersAM = [
-  new UserTop(['BorinWu'], [1094], '80F01'),
-  new UserTop(['concrete'], [1353], '4070B'),
-  new UserTop(['Vegas'], [1026], 'A0B0F'),
-  new UserTop(['pheel'], [1323], '40200'),
-  new UserTop(['MathsDebating'], [1310], 'C050D'),
-  new UserTop(['Zeko'], [1034], 'A0B0F'),
-  new UserTop(['(CH)'], [1019], 'D030F'),
-  new UserTop(['Holo'], [1282], '90504'),
-  new UserTop(['Cookie'], [1004], '1130C'),
-  new UserTop(['Mërk'], [1003], '9190B'),
-  new UserTop(['SorryNotSorry'], [1019], 'A0107'),
-  new UserTop(['Bird'], [1302], '70C0D'),
-  new UserTop(['Mikerodose'], [1121], 'E030D'),
-  new UserTop(['Code'], [1091], 'B0B08'),
-  new UserTop(['Imba'], [1032], 'A040D'),
-  new UserTop(['Chem_Addict'], [1032], 'F0C08'),
-  new UserTop(['Mustard'], [1323], 'F0101'),
-  new UserTop(['PJs'], [1021], 'D0603'),
-  new UserTop(['Huddy'], [1339], '70A07'),
-  new UserTop(['Shroom'], [1250], 'B070D'),
-  new UserTop(['FAFO'], [1021], '40205'),
-  new UserTop(['RougeVifsy'], [1033], 'F0800'),
-  new UserTop(['WaterDog'], [1021], 'B0A03'),
-  new UserTop(['Greed'], [1280], '80101'),
-  new UserTop(['Rubop'], [1027], 'A110F'),
-  new UserTop(['Slovet'], [1027], 'E0806'),
-  new UserTop(['(CH)'], [1019], 'D0406'),
-  new UserTop(['Calmskitzo'], [1226], 'C0101'),
-  new UserTop(['Mocha'], [1121], '8030E'),
-  new UserTop(['Mahoraga'], [1281], 'B0B0E'),
-  new UserTop(['Fuxi'], [1107], '80506'),
-  new UserTop(['8wide'], [1091], 'A0E06'),
-  new UserTop(['Squonk'], [1182], 'E0004'),
-  new UserTop(['SpankA', 'GaryProbably'], [1035], '21A00'),
-  new UserTop(['Morji'], [1004], '2130F'),
-  new UserTop(['Mythic'], [1310], '01A0E'),
-  new UserTop(['(EMPTY)'], [1024], '40001'),
-  new UserTop(['Fara'], [1180], 'E0808'),
-  new UserTop(['Minamoto'], [1117], 'D0401'),
-  new UserTop(['Hya'], [1035], '01306'),
-  new UserTop(['(CH)'], [1019], 'E050C'),
-  new UserTop(['Hollow'], [1024], '5000B'),
-  new UserTop(['Leroy'], [1280], 'C0D07'),
-  new UserTop(['Crusher'], [1474], 'B0F07'),
-  new UserTop(['Coldpulse'], [1461], '11302'),
-  new UserTop(['DrAmALoC248'], [1108], '5040F'),
-  new UserTop(['Azure'], [1339], '40202'),
-  new UserTop(['Bapesta'], [1111], 'A0F01'),
-  new UserTop(['Traub'], [1004], '5020E'),
-  new UserTop(['Barbie'], [1282], 'B0401'),
-  new UserTop(['(?)'], [1032], '6010C'),
-  new UserTop(['Midswift'], [1031], 'F0B0E'),
-  new UserTop(['ShroomieC090C'], [1121], 'C090C'),
-  new UserTop(['KitKat'], [1253], 'E0B0F'),
-  new UserTop(['Feeshy'], [1003], '50305'),
-  new UserTop(['Rafiki'], [1226], 'E0C09'),
-  new UserTop(['JustANobody', 'ZzBLAZE'], [1215], '80F09'),
-  new UserTop(['Joker'], [1254], '61304'),
-  new UserTop(['Harfle'], [1254], '90F0D'),
-  new UserTop(['Bojangles'], [1359], 'B0002'),
-  new UserTop(['Matty'], [1218], '8070F'),
-  new UserTop(['Helly'], [1024], '01B0D'),
-  new UserTop(['Gibby'], [1112], 'C0307'),
-  new UserTop(['Iskander'], [1254], '70C07'),
-  new UserTop(['Famine'], [1117], '8030B'),
-  new UserTop(['Pokey'], [1117], '70506'),
-  new UserTop(['Alteron'], [1013], 'F0E0F'),
-  new UserTop(['Star'], [1002], '61B0F'),
-  new UserTop(['Shroomrix'], [1112], 'B020C'),
-  new UserTop(['Akuma'], [1215], 'B0200'),
-  new UserTop(['pvtmeow'], [1182], 'F0E0C'),
-  new UserTop(['Cheng'], [1003], '21206'),
-  new UserTop(['Archadius'], [1281], 'F0E05'),
-  new UserTop(['Trippy'], [1108], 'A0F04'),
-  new UserTop(['BlackCreme'], [1122], '6070B'),
-  new UserTop(['(JP)'], [1473], 'E0108'),
-  new UserTop(['NormalGuard'], [1003], '41701'),
-  new UserTop(['LowKey'], [1024], '70504'),
-  new UserTop(['Dahlia'], [1403], 'A0807'),
-  new UserTop(['Deathwing'], [1216], 'B0605'),
-
-  new UserTop(['iRETIRED'], [1215], 'F0B00'),
-  new UserTop(['Phemto'], [1281], '50502'),
-  new UserTop(['HUMANKILLER'], [1180], '5130C'),
-  new UserTop(['Oya'], [1321], 'E070D'),
-  new UserTop(['CL4P_TP'], [1404], '6040E'),
-  new UserTop(['Oogie'], [1026], '50F0F'),
-  new UserTop(['Delly'], [1112], 'B0F01'),
-  new UserTop(['DumpsterBacon'], [1170], 'F0E0D'),
-  new UserTop(['johnny'], [1388], '90402'),
-  new UserTop(['Awesome'], [1447], '9020B'),
-  new UserTop(['SnowEmori'], [1112], '50500'),
-  new UserTop(['ShaiLx'], [1019], '21404'),
-  new UserTop(['Diamonds'], [1121], '40207'),
-  new UserTop(['POWER'], [1001], '01B06'),
-  new UserTop(['Deserving'], [1170], '50A06'),
-  new UserTop(['MrBates'], [1369], 'E0E01'),
-  new UserTop(['StepsisLily'], [1421], 'F0A05'),
-  new UserTop(['Poseidon'], [1112], 'C090D'),
-  new UserTop(['Wimm'], [1121], '31E0F'),
-  new UserTop(['WaterFrog'], [1021], '9040C'),
-]
-
-const reiMestre = [
-  new UserMestre(['Doruk'], [11174], '40301'),
-  new UserMestre(['LimRabei'], [11174], '90709'),
-  new UserMestre(['n17'], [11174], 'C0400'),
-  new UserMestre(['n17 2'], [11174], '4040C'),
-  new UserMestre(['KROONOS'], [11220], '50103'),
-  new UserMestre(['Mina'], [11245], '4060A'),
-  // new UserMestre(['PöséjdoN'], [], ''),
-  // new UserMestre(['PhunGuy'], [], ''),
-  // new UserMestre(['GingerKing'], [], ''),
-  // new UserMestre(['Spaceballz'], [], ''),
-  // new UserMestre(['MAGO'], [], ''),
-  // new UserMestre(['KingKurt99'], [], ''),
-  // new UserMestre(['70meMatar'], [], ''),
-  // new UserMestre(['Pwr'], [], ''),
-  // new UserMestre(['Veo3'], [], ''),
-  // new UserMestre(['Nico'], [], ''),
-  // new UserMestre(['RAMØSRP'], [], ''),
-  // new UserMestre(['RegenMyBalls'], [], ''),
-  // new UserMestre(['DEMON☆Pokey'], [], ''),
-  // new UserMestre(['Apollyon'], [], ''),
-  // new UserMestre(['Goats'], [], ''),
-  // new UserMestre(['DarkMagicianGirl'], [], ''),
-  // new UserMestre(['Noobie'], [], ''),
-  // new UserMestre(['Zoeee'], [], ''),
-  // new UserMestre(['Doc'], [], ''),
-  // new UserMestre(['ShroomZillet'], [], ''),
-  // new UserMestre(['DEMON☆Ghost'], [], ''),
-  // new UserMestre(['DaniloLim'], [], ''),
-  // new UserMestre(['POWER'], [], ''),
-  // new UserMestre(['HornyGoat'], [], ''),
-  // new UserMestre(['CheetoBonito'], [], ''),
-  // new UserMestre(['EstarossaRP'], [], ''),
-  // new UserMestre(['SZØRDØ'], [], ''),
-  // new UserMestre(['Soul'], [], ''),
-  // new UserMestre(['Pab'], [], ''),
-  // new UserMestre(['DeathCap'], [], ''),
-  // new UserMestre(['Jessica'], [], ''),
-  // new UserMestre(['Chm'], [], ''),
-  // new UserMestre(['Duvanes'], [], ''),
-  // new UserMestre(['Postmaster'], [], ''),
-  // new UserMestre(['Angel'], [], ''),
-  // new UserMestre(['WhoKnow'], [], ''),
-  // new UserMestre(['n°17'], [], ''),
-  // new UserMestre(['DEMON★PaPa'], [], ''),
-  // new UserMestre(['chobop'], [], ''),
-  // new UserMestre(['MrsHornyGoat'], [], ''),
-  // new UserMestre(['moogen'], [], ''),
-  // new UserMestre(['佛系'], [], ''),
-  // new UserMestre(['iLychéeé'], [], ''),
-  // new UserMestre(['Snoodle'], [], ''),
-]
-
-const honorMention = [
-  new UserHonorMention(['Gogeta'], [11174], '2190B'),
-  new UserHonorMention(['MadHatterVII'], [1321], '60306'),
-  new UserHonorMention(['Raven'], [1019], 'F0002'),
-  new UserHonorMention(['JHotTea'], [1021], '70305'),
-  new UserHonorMention(['HootyHoo', 'BuzzKillington'], [1254], 'D0E0C'),
-  new UserHonorMention(['Remi'], [1173], 'C0F0B'),
-  new UserHonorMention(['Babyboopz'], [1019], '21400'),
-  new UserHonorMention(['Gustavo'], [1107], 'E050E'),
-  new UserHonorMention(['MrsMadHatter'], [1321], 'C070B'),
-  new UserHonorMention(['moose'], [1170], 'E0908'),
-  new UserHonorMention(['Solo'], [1114], '90F06'),
-  new UserHonorMention(['Algor'], [1111], 'E050F'),
-  new UserHonorMention(['Salt'], [1109], 'A0602'),
-  new UserHonorMention(['Callous'], [1321], '90E0B'),
-  new UserHonorMention(['Mojikon'], [1006], '1190A'),
-  new UserHonorMention(['Kobe'], [1031], 'E0D04'),
-  new UserHonorMention(['NrBaruca'], [1107], 'F0E02'),
-  new UserHonorMention(['Balmung'], [1005], '40A0E'),
-  new UserHonorMention(['Dex'], [1008], '4150A'),
-  new UserHonorMention(['mako'], [1182], 'C0A01'),
-  new UserHonorMention(['oobell'], [1218], '8070F'),
-  new UserHonorMention(['shroomdada'], [1005], '80901'),
-  new UserHonorMention(['McFudge'], [1111], '90601'),
-  new UserHonorMention(['Beagle'], [1215], '8000E'),
-  new UserHonorMention(['Kaido'], [1024], '90A02'),
-  new UserHonorMention(['Butgobbler'], [1008], '6000B'),
-  new UserHonorMention(['Benimaru'], [1403], '60607'),
-  new UserHonorMention(['---'], [1250], 'B0302'),
-  new UserHonorMention(['ExpertGuarda'], [1254], '61304'),
-  new UserHonorMention(['Sonic'], [1027], 'B0507'),
-  new UserHonorMention(['Debo'], [1111], 'B0509'),
-  new UserHonorMention(['Bill'], [1003], 'F0207'),
-  new UserHonorMention(['GrimmLord'], [1027], 'B040A'),
-  new UserHonorMention(['InigoMontoya'], [1031], 'A0505'),
-  new UserHonorMention(['uhh'], [1034], 'A0709'),
-  new UserHonorMention(['Halfstep'], [1111], 'B0C08'),
-  new UserHonorMention(['DraunTrager'], [1117], 'C0C0D'),
-  new UserHonorMention(['Darkoslord'], [1122], '60402'),
-  new UserHonorMention(['DixieNormus'], [1217], 'C0A02'),
-  new UserHonorMention(['Asyu'], [1024], '8010A'),
-  new UserHonorMention(['Enchantress'], [1173], '40F00'),
-  new UserHonorMention(['AznRdnk'], [1310], '21105'),
-  new UserHonorMention(['Pavii'], [1216], '11209'),
-  new UserHonorMention(['jett'], [1005], '1140F'),
-  new UserHonorMention(['Beachy'], [1111], '90505'),
-  new UserHonorMention(['ap33zy'], [1109], '9090C'),
-  new UserHonorMention(['Shroomish'], [1339], '90F08'),
-  new UserHonorMention(['Mavis', 'DieEasy'], [1113], '70D05'),
-  new UserHonorMention(['Sura0116'], [1032], '70302'),
-  new UserHonorMention(['Doc'], [1021], 'C0D00'),
-  new UserHonorMention(['Simcraft'], [1302], '0120C'),
-  new UserHonorMention(['GingerKing'], [1033], '50F0A'),
-  new UserHonorMention(['Ry'], [1030], '11C0A'),
-  new UserHonorMention(['Xyz', 'Ray'], [1031], 'D0D0D'),
-  new UserHonorMention(['Brad'], [1282], '8010B'),
-  new UserHonorMention(['captTeemo'], [1280], 'C0B04'),
-  new UserHonorMention(['9030c'], [1182], '9030C'),
-  new UserHonorMention(['Ramos'], [11180], 'E050F'),
-  new UserHonorMention(['O Haio'], [1254], '61304'),
-  new UserHonorMention(['JuraphicSark'], [1002], 'D0D03'),
-  new UserHonorMention(['Sol'], [1304], '90807'),
-  new UserHonorMention(['YodaNBeer'], [1386], '60A03'),
-  new UserHonorMention(['Yasama'], [6119], 'C0603'),
-  new UserHonorMention(['Luxury'], [11179], '7080D'),
-  new UserHonorMention(['Enel'], [1302], '80107'),
-  new UserHonorMention(['Furybound'], [1388], 'C010A'),
-  // new UserTop(['Matty', 'Thing1'], [1218], '11F0B'),
-]
-
-const topUsersEU = [
-  new User(['Dracula'], [36005], 'C0B02'),
-  new User(['RudiRatios'], [33002], '01004'),
-  new User(['StillNoVIP'], [33003], 'D090E'),
-  new User(['Ado'], [33035], 'B010C'),
-  new User(['RGK', 'Nathan'], [36001], '11E02'),
-  new User(['OPMAGAX', 'Guilty'], [30054], '70E02'),
-  new User(['katran'], [30020], '11203'),
-  new User(['Tipo', 'Flope'], [30042], 'C0209'),
-  new User(['Kaffee'], [33015], 'A0F05'),
-  new User(['Massi'], [33002], 'D030D'),
-
-  new User(['Ozan'], [30008], '11304'),
-  new User(['Ado'], [36073], '2110E'),
-  new User(['Nova'], [30054], '2160C'),
-  new User(['Zet'], [30068], '5130F'),
-  new User(['Bongswasser', 'Octopus'], [33002], 'D0B0E'),
-  new User(['Gribnik'], [30054], '2110B'),
-  new User(['MuhQ'], [33036], 'C000D'),
-  new User(['GeeKSlupp'], [30035], '01C04'),
-  new User(['El_Cuddle_Her'], [30077], 'B0009'),
-  new User(['Shaelen'], [30035], '9080C'),
-
-  new User(['JackSporerow', 'Wet Dream'], [33040], '80305'),
-  new User(['Chips'], [30098], 'A0204'),
-  new User(['RM'], [33040], '80608'),
-  new User(['Ralf'], [33061], 'F070B'),
-  new User(['F2P_DritZen'], [36073], 'E0B0B'),
-  new User(['Feixiao'], [33122], '0160F'),
-  new User(['Nouvo'], [30027], 'A1300'),
-  new User(['Jethro', 'Hulago'], [30042], 'A0908'),
-  new User(['Leme'], [33002], '01905'),
-  new User(['Monkey_D_Luffy', 'Wukong'], [36099], 'B0702'),
-
-  new User(['Absuro'], [36133], 'B030B'),
-  new User(['Bluez'], [36073], '70E0A'),
-  new User(['Pikazu'], [33002], '60D0B'),
-  new User(['Horyl'], [36072], '40B05'),
-  new User(['Reigen'], [36120], '50705'),
-  new User(['Lemak76_Pdm'], [30027], 'F040C'),
-  new User(['Catom'], [30053], '2190A'),
-  new User(['Rapunzel'], [33001], '70F08'),
-  new User(['Gelmi'], [33061], 'C0A08'),
-  new User(['Dionysos'], [36099], '51301'),
-
-  new User(['Arkyen'], [30035], 'D020E'),
-  new User(['Preséphone'], [36133], '4030B'),
-  new User(['Baxia'], [33003], 'A0608'),
-  new User(['Shin'], [30077], '50109'),
-  new User(['Dino'], [30001], '11303'),
-  new User(['Ced'], [36072], 'A0002'),
-  new User(['Lazy'], [30054], '50C0C'),
-  new User(['Alice', 'Elemental_XII'], [36001], '71104'),
-  new User(['Kagibi'], [36100], '31005'),
-  new User(['Twisted'], [30068], 'C0500'),
-
-  new User(['Dworky'], [36073], 'F0D08'),
-  new User(['LeRoiBoeuf', 'Achille'], [36099], '8000E'),
-  new User(['Sharleen'], [30020], 'B0307'),
-  new User(['Reezh'], [30053], 'D0B0D'),
-  new User(['Calimero'], [36008], 'A0A08'),
-  new User(['Zoro'], [33036], '60A0A'),
-  new User(['Savage', 'Lipa'], [36073], '80B00'),
-  new User(['Iris'], [30040], '21F09'),
-  new User(['Zdini'], [30020], '11203'),
-  new User(['Fabio'], [33035], 'C0B00'),
-
-  new User(['Gosaitan'], [36008], '21E0C'),
-  new User(['t'], [30027], '4110E'),
-  new User(['Awyy'], [33001], 'A0D05'),
-  new User(['HeraTheCool'], [30008], '21508'),
-  new User(['SnowTime'], [33066], 'D0D02'),
-  new User(['D3MONKILLA'], [30020], '21C02'),
-  new User(['Akho'], [30053], '31107'),
-  new User(['Shadovv'], [33003], '01C06'),
-  new User(['sonix'], [33001], '90506'),
-  new User(['TheShadowMan'], [30001], '31E00'),
-
-  new User(['Poly'], [33066], 'D0C03'),
-  new User(['Caliban'], [33122], 'C0E0C'),
-  new User(['Petit_panda'], [36002], '01007'),
-  new User(['Bao'], [33003], 'B0503'),
-  new User(['Lucas7'], [30065], '3170D'),
-  new User(['Zazou'], [30051], '71D0E'),
-  new User(['Tiny'], [33003], 'E050C'),
-  new User(['Otto'], [30008], '6060C'),
-  new User(['Arsian'], [36005], '0130B'),
-  new User(['Rex'], [30078], '80506'),
-
-  new User(['Konan'], [36072], '11000'),
-  new User(['PJ'], [33002], '90806'),
-  new User(['Lylith'], [36003], '60E07'),
-  new User(['Andi'], [33003], '9070A'),
-  new User(['WickedSoulQ8'], [30064], '4010C'),
-  new User(['Cryptux'], [30008], '40B06'),
-  new User(['Týr', 'WosTheys'], [36001], '2190F'),
-  new User(['Aslang'], [30054], '40C0C'),
-  new User(['Godlike', 'Fubilus'], [33036], '8010B'),
-  new User(['Vasanti'], [30053], 'F090A'),
-
-  new User(['Xel'], [30097], '90A00'),
-  new User(['7000B', '15xRAE'], [33015], '7000B'),
-  new User(['FLASHBIRD'], [33036], 'A050B'),
-  new User(['Scarlet'], [30078], '50400'),
-  new User(['ShanksLeRoux'], [30097], 'A0C08'),
-  new User(['Freya'], [36132], '4090B'),
-  new User(['freedom'], [33117], '0170D'),
-  // new User(['Etozhetsve', 'NA_CWE'], [30098], 'C)B00'),
-  new User(['F2P_Holoflux'], [33103], '50905'),
-  new User(['GoblinSlayer'], [33061], '4010F'),
-]
-
-const users = [
-  ...friends,
-  ...topUsersAM,
-  ...topUsersEU,
-  ...reiMestre,
-  ...honorMention,
-].map(user => {
-  const result = user
-  result.name = [result.id, result.name].filter(e => e).join(' :: ')
-  return result
-})
-
-export { merges, users as usersDB }
