+++ lom/js/db-users-elite.js
@@ -1,0 +1,54 @@
+import { UserElite } from './classes.js'
+
+export const elites = [
+  new UserElite(['Doruk'], [11174], '40301'),
+  new UserElite(['LimRabei'], [11174], '90709'),
+  new UserElite(['n17'], [11174], 'C0400'),
+  new UserElite(['n17 2'], [11174], '4040C'),
+  new UserElite(['KROONOS'], [11220], '50103'),
+  new UserElite(['Mina'], [11245], '4060A'),
+  // new UserElite(['PöséjdoN'], [], ''),
+  // new UserElite(['PhunGuy'], [], ''),
+  // new UserElite(['GingerKing'], [], ''),
+  // new UserElite(['Spaceballz'], [], ''),
+  // new UserElite(['MAGO'], [], ''),
+  // new UserElite(['KingKurt99'], [], ''),
+  // new UserElite(['70meMatar'], [], ''),
+  // new UserElite(['Pwr'], [], ''),
+  // new UserElite(['Veo3'], [], ''),
+  // new UserElite(['Nico'], [], ''),
+  // new UserElite(['RAMØSRP'], [], ''),
+  // new UserElite(['RegenMyBalls'], [], ''),
+  // new UserElite(['DEMON☆Pokey'], [], ''),
+  // new UserElite(['Apollyon'], [], ''),
+  // new UserElite(['Goats'], [], ''),
+  // new UserElite(['DarkMagicianGirl'], [], ''),
+  // new UserElite(['Noobie'], [], ''),
+  // new UserElite(['Zoeee'], [], ''),
+  // new UserElite(['Doc'], [], ''),
+  // new UserElite(['ShroomZillet'], [], ''),
+  // new UserElite(['DEMON☆Ghost'], [], ''),
+  // new UserElite(['DaniloLim'], [], ''),
+  // new UserElite(['POWER'], [], ''),
+  // new UserElite(['HornyGoat'], [], ''),
+  // new UserElite(['CheetoBonito'], [], ''),
+  // new UserElite(['EstarossaRP'], [], ''),
+  // new UserElite(['SZØRDØ'], [], ''),
+  // new UserElite(['Soul'], [], ''),
+  // new UserElite(['Pab'], [], ''),
+  // new UserElite(['DeathCap'], [], ''),
+  // new UserElite(['Jessica'], [], ''),
+  // new UserElite(['Chm'], [], ''),
+  // new UserElite(['Duvanes'], [], ''),
+  // new UserElite(['Postmaster'], [], ''),
+  // new UserElite(['Angel'], [], ''),
+  // new UserElite(['WhoKnow'], [], ''),
+  // new UserElite(['n°17'], [], ''),
+  // new UserElite(['DEMON★PaPa'], [], ''),
+  // new UserElite(['chobop'], [], ''),
+  // new UserElite(['MrsHornyGoat'], [], ''),
+  // new UserElite(['moogen'], [], ''),
+  // new UserElite(['佛系'], [], ''),
+  // new UserElite(['iLychéeé'], [], ''),
+  // new UserElite(['Snoodle'], [], ''),
+]
