+++ lom/js/db-users-friends.js
@@ -1,0 +1,19 @@
+import { User } from './classes.js'
+
+export const friends = [
+  new User(['DKiingG7'], [11042], null, ['friend']),
+  new User(['AhkTaar'], [30001], null, ['friend']),
+  // new User(['Talisa'], [1310, 6103], null, ['friend']),
+  // new User(['Nico'], [1310, 6103], null, ['friend']),
+  // new User(['Doruk'], [11174], null, ['friend']),
+  // new User(['Koturno'], [1388, 11220, 11222], null, ['friend']),
+  // new User(['Neo'], [11256], null, ['friend']),
+  // new User(['Diego'], [11257], null, ['friend']),
+  // new User(['Gustavo'], [11310], null, ['friend']),
+  // new User(['Murilo'], [11300], null, ['friend']),
+  // new User(['Cleo'], [11323], null, ['friend']),
+  // new User(['Max'], [11323], null, ['friend']),
+  // new User(['Mandy'], [11323], null, ['friend']),
+  // new User(['Magneto'], [11323], null, ['friend']),
+  // new User(['Ipek'], [36264], null, ['friend']),
+]
