+++ lom/js/db-users-honor.js
@@ -1,0 +1,71 @@
+import { UserHonorMention } from './classes.js'
+
+export const honorMention = [
+  new UserHonorMention(['Gogeta'], [11174], '2190B'),
+  new UserHonorMention(['MadHatterVII'], [1321], '60306'),
+  new UserHonorMention(['Raven'], [1019], 'F0002'),
+  new UserHonorMention(['JHotTea'], [1021], '70305'),
+  new UserHonorMention(['HootyHoo', 'BuzzKillington'], [1254], 'D0E0C'),
+  new UserHonorMention(['Remi'], [1173], 'C0F0B'),
+  new UserHonorMention(['Babyboopz'], [1019], '21400'),
+  new UserHonorMention(['Gustavo'], [1107], 'E050E'),
+  new UserHonorMention(['MrsMadHatter'], [1321], 'C070B'),
+  new UserHonorMention(['moose'], [1170], 'E0908'),
+  new UserHonorMention(['Solo'], [1114], '90F06'),
+  new UserHonorMention(['Algor'], [1111], 'E050F'),
+  new UserHonorMention(['Salt'], [1109], 'A0602'),
+  new UserHonorMention(['Callous'], [1321], '90E0B'),
+  new UserHonorMention(['Mojikon'], [1006], '1190A'),
+  new UserHonorMention(['Kobe'], [1031], 'E0D04'),
+  new UserHonorMention(['NrBaruca'], [1107], 'F0E02'),
+  new UserHonorMention(['Balmung'], [1005], '40A0E'),
+  new UserHonorMention(['Dex'], [1008], '4150A'),
+  new UserHonorMention(['mako'], [1182], 'C0A01'),
+  new UserHonorMention(['oobell'], [1218], '8070F'),
+  new UserHonorMention(['shroomdada'], [1005], '80901'),
+  new UserHonorMention(['McFudge'], [1111], '90601'),
+  new UserHonorMention(['Beagle'], [1215], '8000E'),
+  new UserHonorMention(['Kaido'], [1024], '90A02'),
+  new UserHonorMention(['Butgobbler'], [1008], '6000B'),
+  new UserHonorMention(['Benimaru'], [1403], '60607'),
+  new UserHonorMention(['---'], [1250], 'B0302'),
+  new UserHonorMention(['ExpertGuarda'], [1254], '61304'),
+  new UserHonorMention(['Sonic'], [1027], 'B0507'),
+  new UserHonorMention(['Debo'], [1111], 'B0509'),
+  new UserHonorMention(['Bill'], [1003], 'F0207'),
+  new UserHonorMention(['GrimmLord'], [1027], 'B040A'),
+  new UserHonorMention(['InigoMontoya'], [1031], 'A0505'),
+  new UserHonorMention(['uhh'], [1034], 'A0709'),
+  new UserHonorMention(['Halfstep'], [1111], 'B0C08'),
+  new UserHonorMention(['DraunTrager'], [1117], 'C0C0D'),
+  new UserHonorMention(['Darkoslord'], [1122], '60402'),
+  new UserHonorMention(['DixieNormus'], [1217], 'C0A02'),
+  new UserHonorMention(['Asyu'], [1024], '8010A'),
+  new UserHonorMention(['Enchantress'], [1173], '40F00'),
+  new UserHonorMention(['AznRdnk'], [1310], '21105'),
+  new UserHonorMention(['Pavii'], [1216], '11209'),
+  new UserHonorMention(['jett'], [1005], '1140F'),
+  new UserHonorMention(['Beachy'], [1111], '90505'),
+  new UserHonorMention(['ap33zy'], [1109], '9090C'),
+  new UserHonorMention(['Shroomish'], [1339], '90F08'),
+  new UserHonorMention(['Mavis', 'DieEasy'], [1113], '70D05'),
+  new UserHonorMention(['Sura0116'], [1032], '70302'),
+  new UserHonorMention(['Doc'], [1021], 'C0D00'),
+  new UserHonorMention(['Simcraft'], [1302], '0120C'),
+  new UserHonorMention(['GingerKing'], [1033], '50F0A'),
+  new UserHonorMention(['Ry'], [1030], '11C0A'),
+  new UserHonorMention(['Xyz', 'Ray'], [1031], 'D0D0D'),
+  new UserHonorMention(['Brad'], [1282], '8010B'),
+  new UserHonorMention(['captTeemo'], [1280], 'C0B04'),
+  new UserHonorMention(['9030c'], [1182], '9030C'),
+  new UserHonorMention(['Ramos'], [11180], 'E050F'),
+  new UserHonorMention(['O Haio'], [1254], '61304'),
+  new UserHonorMention(['JuraphicSark'], [1002], 'D0D03'),
+  new UserHonorMention(['Sol'], [1304], '90807'),
+  new UserHonorMention(['YodaNBeer'], [1386], '60A03'),
+  new UserHonorMention(['Yasama'], [6119], 'C0603'),
+  new UserHonorMention(['Luxury'], [11179], '7080D'),
+  new UserHonorMention(['Enel'], [1302], '80107'),
+  new UserHonorMention(['Furybound'], [1388], 'C010A'),
+  // new UserTop(['Matty', 'Thing1'], [1218], '11F0B'),
+]
