+++ lom/js/db-users-top-am.js
@@ -1,0 +1,105 @@
+import { UserTop } from './classes.js'
+
+export const topAM = [
+  new UserTop(['BorinWu'], [1094], '80F01'),
+  new UserTop(['concrete'], [1353], '4070B'),
+  new UserTop(['Vegas'], [1026], 'A0B0F'),
+  new UserTop(['pheel'], [1323], '40200'),
+  new UserTop(['MathsDebating'], [1310], 'C050D'),
+  new UserTop(['Zeko'], [1034], 'A0B0F'),
+  new UserTop(['(CH)'], [1019], 'D030F'),
+  new UserTop(['Holo'], [1282], '90504'),
+  new UserTop(['Cookie'], [1004], '1130C'),
+  new UserTop(['Mërk'], [1003], '9190B'),
+  new UserTop(['SorryNotSorry'], [1019], 'A0107'),
+  new UserTop(['Bird'], [1302], '70C0D'),
+  new UserTop(['Mikerodose'], [1121], 'E030D'),
+  new UserTop(['Code'], [1091], 'B0B08'),
+  new UserTop(['Imba'], [1032], 'A040D'),
+  new UserTop(['Chem_Addict'], [1032], 'F0C08'),
+  new UserTop(['Mustard'], [1323], 'F0101'),
+  new UserTop(['PJs'], [1021], 'D0603'),
+  new UserTop(['Huddy'], [1339], '70A07'),
+  new UserTop(['Shroom'], [1250], 'B070D'),
+  new UserTop(['FAFO'], [1021], '40205'),
+  new UserTop(['RougeVifsy'], [1033], 'F0800'),
+  new UserTop(['WaterDog'], [1021], 'B0A03'),
+  new UserTop(['Greed'], [1280], '80101'),
+  new UserTop(['Rubop'], [1027], 'A110F'),
+  new UserTop(['Slovet'], [1027], 'E0806'),
+  new UserTop(['(CH)'], [1019], 'D0406'),
+  new UserTop(['Calmskitzo'], [1226], 'C0101'),
+  new UserTop(['Mocha'], [1121], '8030E'),
+  new UserTop(['Mahoraga'], [1281], 'B0B0E'),
+  new UserTop(['Fuxi'], [1107], '80506'),
+  new UserTop(['8wide'], [1091], 'A0E06'),
+  new UserTop(['Squonk'], [1182], 'E0004'),
+  new UserTop(['SpankA', 'GaryProbably'], [1035], '21A00'),
+  new UserTop(['Morji'], [1004], '2130F'),
+  new UserTop(['Mythic'], [1310], '01A0E'),
+  new UserTop(['(EMPTY)'], [1024], '40001'),
+  new UserTop(['Fara'], [1180], 'E0808'),
+  new UserTop(['Minamoto'], [1117], 'D0401'),
+  new UserTop(['Hya'], [1035], '01306'),
+  new UserTop(['(CH)'], [1019], 'E050C'),
+  new UserTop(['Hollow'], [1024], '5000B'),
+  new UserTop(['Leroy'], [1280], 'C0D07'),
+  new UserTop(['Crusher'], [1474], 'B0F07'),
+  new UserTop(['Coldpulse'], [1461], '11302'),
+  new UserTop(['DrAmALoC248'], [1108], '5040F'),
+  new UserTop(['Azure'], [1339], '40202'),
+  new UserTop(['Bapesta'], [1111], 'A0F01'),
+  new UserTop(['Traub'], [1004], '5020E'),
+  new UserTop(['Barbie'], [1282], 'B0401'),
+  new UserTop(['(?)'], [1032], '6010C'),
+  new UserTop(['Midswift'], [1031], 'F0B0E'),
+  new UserTop(['ShroomieC090C'], [1121], 'C090C'),
+  new UserTop(['KitKat'], [1253], 'E0B0F'),
+  new UserTop(['Feeshy'], [1003], '50305'),
+  new UserTop(['Rafiki'], [1226], 'E0C09'),
+  new UserTop(['JustANobody', 'ZzBLAZE'], [1215], '80F09'),
+  new UserTop(['Joker'], [1254], '61304'),
+  new UserTop(['Harfle'], [1254], '90F0D'),
+  new UserTop(['Bojangles'], [1359], 'B0002'),
+  new UserTop(['Matty'], [1218], '8070F'),
+  new UserTop(['Helly'], [1024], '01B0D'),
+  new UserTop(['Gibby'], [1112], 'C0307'),
+  new UserTop(['Iskander'], [1254], '70C07'),
+  new UserTop(['Famine'], [1117], '8030B'),
+  new UserTop(['Pokey'], [1117], '70506'),
+  new UserTop(['Alteron'], [1013], 'F0E0F'),
+  new UserTop(['Star'], [1002], '61B0F'),
+  new UserTop(['Shroomrix'], [1112], 'B020C'),
+  new UserTop(['Akuma'], [1215], 'B0200'),
+  new UserTop(['pvtmeow'], [1182], 'F0E0C'),
+  new UserTop(['Cheng'], [1003], '21206'),
+  new UserTop(['Archadius'], [1281], 'F0E05'),
+  new UserTop(['Trippy'], [1108], 'A0F04'),
+  new UserTop(['BlackCreme'], [1122], '6070B'),
+  new UserTop(['(JP)'], [1473], 'E0108'),
+  new UserTop(['NormalGuard'], [1003], '41701'),
+  new UserTop(['LowKey'], [1024], '70504'),
+  new UserTop(['Dahlia'], [1403], 'A0807'),
+  new UserTop(['Deathwing'], [1216], 'B0605'),
+
+  new UserTop(['iRETIRED'], [1215], 'F0B00'),
+  new UserTop(['Phemto'], [1281], '50502'),
+  new UserTop(['HUMANKILLER'], [1180], '5130C'),
+  new UserTop(['Oya'], [1321], 'E070D'),
+  new UserTop(['CL4P_TP'], [1404], '6040E'),
+  new UserTop(['Oogie'], [1026], '50F0F'),
+  new UserTop(['Delly'], [1112], 'B0F01'),
+  new UserTop(['DumpsterBacon'], [1170], 'F0E0D'),
+  new UserTop(['johnny'], [1388], '90402'),
+  new UserTop(['Awesome'], [1447], '9020B'),
+  new UserTop(['SnowEmori'], [1112], '50500'),
+  new UserTop(['ShaiLx'], [1019], '21404'),
+  new UserTop(['Diamonds'], [1121], '40207'),
+  new UserTop(['POWER'], [1001], '01B06'),
+  new UserTop(['Deserving'], [1170], '50A06'),
+  new UserTop(['MrBates'], [1369], 'E0E01'),
+  new UserTop(['StepsisLily'], [1421], 'F0A05'),
+  new UserTop(['Poseidon'], [1112], 'C090D'),
+  new UserTop(['Wimm'], [1121], '31E0F'),
+  new UserTop(['WaterFrog'], [1021], '9040C'),
+]
