+++ lom/js/db-users-top-eu.js
@@ -1,0 +1,113 @@
+import { UserTop } from './classes.js'
+
+export const topEU = [
+  new UserTop(['Dracula'], [36005], 'C0B02'),
+  new UserTop(['RudiRatios'], [33002], '01004'),
+  new UserTop(['StillNoVIP'], [33003], 'D090E'),
+  new UserTop(['Ado'], [33035], 'B010C'),
+  new UserTop(['RGK', 'Nathan'], [36001], '11E02'),
+  new UserTop(['OPMAGAX', 'Guilty'], [30054], '70E02'),
+  new UserTop(['katran'], [30020], '11203'),
+  new UserTop(['Tipo', 'Flope'], [30042], 'C0209'),
+  new UserTop(['Kaffee'], [33015], 'A0F05'),
+  new UserTop(['Massi'], [33002], 'D030D'),
+
+  new UserTop(['Ozan'], [30008], '11304'),
+  new UserTop(['Ado'], [36073], '2110E'),
+  new UserTop(['Nova'], [30054], '2160C'),
+  new UserTop(['Zet'], [30068], '5130F'),
+  new UserTop(['Bongswasser', 'Octopus'], [33002], 'D0B0E'),
+  new UserTop(['Gribnik'], [30054], '2110B'),
+  new UserTop(['MuhQ'], [33036], 'C000D'),
+  new UserTop(['GeeKSlupp'], [30035], '01C04'),
+  new UserTop(['El_Cuddle_Her'], [30077], 'B0009'),
+  new UserTop(['Shaelen'], [30035], '9080C'),
+
+  new UserTop(['JackSporerow', 'Wet Dream'], [33040], '80305'),
+  new UserTop(['Chips'], [30098], 'A0204'),
+  new UserTop(['RM'], [33040], '80608'),
+  new UserTop(['Ralf'], [33061], 'F070B'),
+  new UserTop(['F2P_DritZen'], [36073], 'E0B0B'),
+  new UserTop(['Feixiao'], [33122], '0160F'),
+  new UserTop(['Nouvo'], [30027], 'A1300'),
+  new UserTop(['Jethro', 'Hulago'], [30042], 'A0908'),
+  new UserTop(['Leme'], [33002], '01905'),
+  new UserTop(['Monkey_D_Luffy', 'Wukong'], [36099], 'B0702'),
+
+  new UserTop(['Absuro'], [36133], 'B030B'),
+  new UserTop(['Bluez'], [36073], '70E0A'),
+  new UserTop(['Pikazu'], [33002], '60D0B'),
+  new UserTop(['Horyl'], [36072], '40B05'),
+  new UserTop(['Reigen'], [36120], '50705'),
+  new UserTop(['Lemak76_Pdm'], [30027], 'F040C'),
+  new UserTop(['Catom'], [30053], '2190A'),
+  new UserTop(['Rapunzel'], [33001], '70F08'),
+  new UserTop(['Gelmi'], [33061], 'C0A08'),
+  new UserTop(['Dionysos'], [36099], '51301'),
+
+  new UserTop(['Arkyen'], [30035], 'D020E'),
+  new UserTop(['Preséphone'], [36133], '4030B'),
+  new UserTop(['Baxia'], [33003], 'A0608'),
+  new UserTop(['Shin'], [30077], '50109'),
+  new UserTop(['Dino'], [30001], '11303'),
+  new UserTop(['Ced'], [36072], 'A0002'),
+  new UserTop(['Lazy'], [30054], '50C0C'),
+  new UserTop(['Alice', 'Elemental_XII'], [36001], '71104'),
+  new UserTop(['Kagibi'], [36100], '31005'),
+  new UserTop(['Twisted'], [30068], 'C0500'),
+
+  new UserTop(['Dworky'], [36073], 'F0D08'),
+  new UserTop(['LeRoiBoeuf', 'Achille'], [36099], '8000E'),
+  new UserTop(['Sharleen'], [30020], 'B0307'),
+  new UserTop(['Reezh'], [30053], 'D0B0D'),
+  new UserTop(['Calimero'], [36008], 'A0A08'),
+  new UserTop(['Zoro'], [33036], '60A0A'),
+  new UserTop(['Savage', 'Lipa'], [36073], '80B00'),
+  new UserTop(['Iris'], [30040], '21F09'),
+  new UserTop(['Zdini'], [30020], '11203'),
+  new UserTop(['Fabio'], [33035], 'C0B00'),
+
+  new UserTop(['Gosaitan'], [36008], '21E0C'),
+  new UserTop(['t'], [30027], '4110E'),
+  new UserTop(['Awyy'], [33001], 'A0D05'),
+  new UserTop(['HeraTheCool'], [30008], '21508'),
+  new UserTop(['SnowTime'], [33066], 'D0D02'),
+  new UserTop(['D3MONKILLA'], [30020], '21C02'),
+  new UserTop(['Akho'], [30053], '31107'),
+  new UserTop(['Shadovv'], [33003], '01C06'),
+  new UserTop(['sonix'], [33001], '90506'),
+  new UserTop(['TheShadowMan'], [30001], '31E00'),
+
+  new UserTop(['Poly'], [33066], 'D0C03'),
+  new UserTop(['Caliban'], [33122], 'C0E0C'),
+  new UserTop(['Petit_panda'], [36002], '01007'),
+  new UserTop(['Bao'], [33003], 'B0503'),
+  new UserTop(['Lucas7'], [30065], '3170D'),
+  new UserTop(['Zazou'], [30051], '71D0E'),
+  new UserTop(['Tiny'], [33003], 'E050C'),
+  new UserTop(['Otto'], [30008], '6060C'),
+  new UserTop(['Arsian'], [36005], '0130B'),
+  new UserTop(['Rex'], [30078], '80506'),
+
+  new UserTop(['Konan'], [36072], '11000'),
+  new UserTop(['PJ'], [33002], '90806'),
+  new UserTop(['Lylith'], [36003], '60E07'),
+  new UserTop(['Andi'], [33003], '9070A'),
+  new UserTop(['WickedSoulQ8'], [30064], '4010C'),
+  new UserTop(['Cryptux'], [30008], '40B06'),
+  new UserTop(['Týr', 'WosTheys'], [36001], '2190F'),
+  new UserTop(['Aslang'], [30054], '40C0C'),
+  new UserTop(['Godlike', 'Fubilus'], [33036], '8010B'),
+  new UserTop(['Vasanti'], [30053], 'F090A'),
+
+  new UserTop(['Xel'], [30097], '90A00'),
+  new UserTop(['7000B', '15xRAE'], [33015], '7000B'),
+  new UserTop(['FLASHBIRD'], [33036], 'A050B'),
+  new UserTop(['Scarlet'], [30078], '50400'),
+  new UserTop(['ShanksLeRoux'], [30097], 'A0C08'),
+  new UserTop(['Freya'], [36132], '4090B'),
+  new UserTop(['freedom'], [33117], '0170D'),
+  // new UserTop(['Etozhetsve', 'NA_CWE'], [30098], 'C)B00'),
+  new UserTop(['F2P_Holoflux'], [33103], '50905'),
+  new UserTop(['GoblinSlayer'], [33061], '4010F'),
+]
