+++ lom/js/simple.js
@@ -1,50 +1,35 @@
 import '../../_global/js/index.js'
-import { merges } from './db.js'
+import { merges } from './db-merges.js'
 import { getCountryCode, getTooltip } from './utils.js'
 import { getPrefix, createNode } from '../../_global/js/global.js'
+import { filters } from './comp-filters.js'
 import './users.js'
 
 const template = `
   <nn-caja padding="4">
-		<div class="nav">
-			<div class="controllers">
-				<button id="all">ALL</button>
-				<button id="amen">AMEN</button>
-				<button id="es">ES</button>
-				<button id="espt">ESPT</button>
-				<button id="pt">PT</button>
-				<button id="euen">EUEN</button>
-				<button id="mush">MUSH</button>
-				<button id="de">DE</button>
-				<button id="fr">FR</button>
-				<button id="me">ME</button>
-				<button id="tr">TR</button>
-				<button id="ru">RU</button>
-				<div id="honor">MENTION</div>
-				<div id="mestre">ELITE</div>
-				<button id="top">TOP 100</button>
-			</div>
-		</div>
+    ${filters}
     <div id="merged-list" class="merged-table"></div>
   </nn-caja>
 `
 const data = {
   attrs: [],
   language: 'all',
-  justTops: false,
-  allMerges: Object.entries(merges).map(([key, values]) => {
-    const servers = Object.entries(values).map(([mergeKey, mergeValues]) => {
+  statusFilter: undefined,
+  allMerges: Object.entries(merges)
+    .map(([key, values]) => {
+      const servers = Object.entries(values).map(([mergeKey, mergeValues]) => {
+        return {
+          key: getCountryCode(mergeKey),
+          values: mergeValues.map(val => getCountryCode(val)),
+        }
+      })
+
       return {
-        key: getCountryCode(mergeKey),
-        values: mergeValues.map(val => getCountryCode(val)),
+        date: key,
+        servers,
       }
     })
-
-    return {
-      date: key,
-      servers,
-    }
-  }).sort((a, b) => new Date(b.date) - new Date(a.date)),
+    .sort((a, b) => new Date(b.date) - new Date(a.date)),
 }
 
 class Simple extends HTMLElement {
@@ -69,15 +54,27 @@ class Simple extends HTMLElement {
     ].forEach(lang => {
       document.getElementById(lang).addEventListener('click', () => {
         data.language = lang
-        data.justTops = false
-        this.generateTable(data.language, data.justTops)
+        data.statusFilter = undefined
+        this.generateTable(data.language, data.statusFilter)
       })
+    })
+
+    document.getElementById('honor').addEventListener('click', () => {
+      data.language = 'all'
+      data.statusFilter = 'honor'
+      this.generateTable(data.language, data.statusFilter)
+    })
+
+    document.getElementById('elite').addEventListener('click', () => {
+      data.language = 'all'
+      data.statusFilter = 'elite'
+      this.generateTable(data.language, data.statusFilter)
     })
 
     document.getElementById('top').addEventListener('click', () => {
       data.language = 'all'
-      data.justTops = true
-      this.generateTable(data.language, data.justTops)
+      data.statusFilter = 'top'
+      this.generateTable(data.language, data.statusFilter)
     })
   }
 
@@ -116,12 +113,16 @@ class Simple extends HTMLElement {
             server.key.code === filterBy ||
             server.values.some(val => val.code === filterBy)
         )
-      } else if (data.justTops) {
+      } else if (data.statusFilter) {
         localServers = merge.servers.filter(
           server =>
-            server.key.users.some(user => user?.group?.includes('top')) ||
+            server.key.users.some(user =>
+              user?.group?.includes(data.statusFilter)
+            ) ||
             server.values.some(server =>
-              server.users.some(user => user?.group?.includes('top'))
+              server.users.some(user =>
+                user?.group?.includes(data.statusFilter)
+              )
             )
         )
       } else {
