+++ lom/js/utils.js
@@ -1,4 +1,4 @@
-import { usersDB } from './db.js'
+import { usersDB } from './db-users.js'
 import { countryCodes } from './country-codes.js'
 
 function getTooltip(item, group = 0) {
